refactor: Project directory reorganization
