+++ app/src/main/java/tech/capullo/radio/espoti/AndroidNativeDecoderInitializer.kt
@@ -1,8 +1,7 @@
-package tech.capullo.radio
+package tech.capullo.radio.espoti
 
 import android.content.Context
 import androidx.startup.Initializer
-import tech.capullo.radio.data.AndroidNativeDecoder
 import xyz.gianlu.librespot.audio.decoders.Decoders
 import xyz.gianlu.librespot.audio.format.SuperAudioFormat
